+++ openlibrary/plugins/openlibrary/api.py
@@ -658,13 +658,10 @@ class hide_banner(delegate.page):
 
     def POST(self):
         data = json.loads(web.data())
-        user = accounts.get_current_user()
 
-        user.save_preferences({'hidden-banner': data['storage-key']})
+        # Set truthy cookie that expires in 30 days:
+        web.setcookie(data['cookie-name'], '1', expires=(3600 * 24 * 30))
 
-        def response(msg, status="success"):
-            return delegate.RawText(
-                json.dumps({status: msg}), content_type="application/json"
-            )
-
-        return response('Preference saved')
+        return delegate.RawText(
+            json.dumps({'success': 'Preference saved'}), content_type="application/json"
+        )
